12_Challenge_Finantial_Agent/pyproject.toml
[project]
name = "12-challenge-finantial-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.49.1",
    "langchain>=1.0.7",
    "langchain-ollama>=1.0.0",
    "pyyaml>=6.0.3",
]
